refactor: app wave-15 Slice C Flame splits and 300-line ratchet
Replace GameMapAreaStateLogic forwarder with topic barrels, split fleet draft and region-map modules over 300 lines, and lower repo.app_flame_lib_physical_file_size to 300.
