+++ SPEC/program/repo-lint.md
@@ -163,7 +163,7 @@
 | `tool/check_orders_test_support_loc.dart` | Physical-line ceilings for `packages/colonizethis_orders/test/orders/support/**/*.dart` (wave-5 support AC ≤13,950) and package `test/**/*.dart` (wave-8 package AC ≤16,000 after support densify; military counsel ranking scenarios fit within wave-8 ceiling; Refs #4109, #4246, #4262, #4281, #4282, #4317, #4307); rule `repo.orders_test_support_loc` |
 | `tool/check_orders_test_shorthand_file_size.dart` | Physical-line ceiling (**180**) for every `*expectation_shorthand.dart` under `packages/colonizethis_orders/test/orders/support/**` (Refs #4109 wave 5); rule `repo.orders_test_shorthand_file_size` |
 | `tool/check_orders_lib_source_file_size.dart` | Ratchet ceiling for `packages/colonizethis_orders/lib/**` physical lines per file (wave-8 ≤320; empty shrink-only grandfather; generated suffixes excluded); rule `repo.orders_lib_source_file_size` (Refs #4246, #4317) |
-| `tool/check_app_flame_lib_physical_file_size.dart` | Ratchet ceiling for `app/lib/features/game/flame/**` physical lines per file (wave-10 ≤400; shrink-only grandfather allowlist for region-map render modules and near-cap province-action states during transition; generated suffixes excluded); rule `repo.app_flame_lib_physical_file_size` (Refs #4183) |
+| `tool/check_app_flame_lib_physical_file_size.dart` | Ratchet ceiling for `app/lib/features/game/flame/**` physical lines per file (wave-15 ≤300; shrink-only grandfather allowlist empty after slice C; generated suffixes excluded); rule `repo.app_flame_lib_physical_file_size` (Refs #4352) |
 | `tool/check_app_features_screens_lib_physical_file_size.dart` | Ratchet ceiling for `app/lib/features/game/screens/**` physical lines per file (wave-14 ≤400; shrink-only grandfather during transition; generated suffixes excluded); rule `repo.app_features_screens_lib_physical_file_size` (Refs #4305 Slice B) |
 | `tool/check_orders_test_support_raw_game.dart` | Ceiling on `return Game(` / `=> Game(` under `test/orders/support/**` excluding `common/` (wave-4 migration ratchet; Refs #3971); rule `repo.orders_test_support_raw_game` |
 | `tool/check_orders_scenario_table_runner.dart` | Prefer scenario-table runners for new long imperative `test()` bodies under `packages/colonizethis_orders/test/**`; after wave-3 slice 13, baseline allow-all is off and remaining imperative suites are listed in `ordersPreferScenarioTablesAllowlist` / `REFACTOR_TRACE.md` (Refs #3949); rule `repo.orders_test_prefer_scenario_tables` |
