+++ app/lib/features/game/flame/map_state/game_map_area_selection.dart
@@ -81,7 +81,7 @@ mixin GameMapAreaSelection on ConsumerState<GameMapArea>, GameMapAreaStateBase {
     final view = buildPlayerView(game, topology, mapPlayerId);
     final workTarget = workTargetSelection!.workTarget;
     cachedValidTileKeys =
-        GameMapAreaStateLogic.resolveValidTileKeysForCivilianWorkSelection(
+        GameMapAreaStateLogicWorkTargets.resolveValidTileKeysForCivilianWorkSelection(
           workTarget: workTarget,
           workTargetSelectionCache: workTargetSelectionCache,
           humanPlayerId: mapPlayerId,
@@ -156,7 +156,7 @@ mixin GameMapAreaSelection on ConsumerState<GameMapArea>, GameMapAreaStateBase {
     final sel = workTargetSelection;
     if (sel == null) return;
     final target = sel.workTarget;
-    final targetTileKey = GameMapAreaStateLogic.translateWorkTargetTileKey(
+    final targetTileKey = GameMapAreaStateLogicShell.translateWorkTargetTileKey(
       tileKey: tileKey,
       workTarget: target,
     );
@@ -169,7 +169,7 @@ mixin GameMapAreaSelection on ConsumerState<GameMapArea>, GameMapAreaStateBase {
     ref
         .read(currentOrdersProvider.notifier)
         .replaceAll(
-          GameMapAreaStateLogic.addHumanWorkOrder(
+          GameMapAreaStateLogicWorkTargets.addHumanWorkOrder(
             orders: orders,
             humanPlayerId: mapPlayerId,
             workOrder: workOrder,
@@ -177,7 +177,7 @@ mixin GameMapAreaSelection on ConsumerState<GameMapArea>, GameMapAreaStateBase {
         );
     setState(() {
       selectedCivilianTileKey =
-          GameMapAreaStateLogic.selectionAfterWorkAssignment(
+          GameMapAreaStateLogicWorkTargets.selectionAfterWorkAssignment(
             currentSelectedCivilianTileKey: selectedCivilianTileKey,
             assignedTileKey: targetTileKey,
           );
