+++ app/lib/features/game/flame/map_state/game_map_area_state_logic.dart
@@ -1,382 +1,9 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
-
-import 'package:colonizethis_map/colonizethis_map.dart';
-import 'package:colonizethis_models/colonizethis_models.dart' as ct_models;
-
-import 'game_map_area_province_action_states.dart';
-import 'game_map_area_state_logic_draft_projection.dart';
-import 'game_map_area_state_logic_province_actions.dart';
-import 'game_map_area_state_logic_shell.dart';
-import 'game_map_area_state_logic_work_targets.dart';
-
+/// Topic barrels for `GameMapArea` state translation helpers.
+///
+/// Wave-15 (#4352 Slice C): this file is the stable export surface; call sites
+/// use topic classes directly (`GameMapAreaStateLogicShell`, …).
+export 'game_map_area_province_action_states.dart';
 export 'game_map_area_state_logic_draft_projection.dart';
 export 'game_map_area_state_logic_province_actions.dart';
 export 'game_map_area_state_logic_shell.dart';
 export 'game_map_area_state_logic_work_targets.dart';
-import 'package:colonizethis_world/colonizethis_world.dart';
-import 'package:colonizethis_logic/ai_api.dart';
-import 'package:colonizethis_orders/colonizethis_orders.dart';
-
-/// Pure-ish helpers for `GameMapArea` state translation.
-///
-/// Splits per `SPEC/program/dart-file-non-comment-line-size.md` and #2575
-/// work item 11. Detailed projection and province-action pipelines live in
-/// dedicated modules; this class keeps the public entry points used by the
-/// `GameMapArea` widget, tests, and the order-suggestions SPEC pointer
-/// (`SPEC/program/order-suggestions.md` § Authoritative pipeline).
-///
-/// Wave-5 (#4018): calls shell / work-target / draft / province modules
-/// directly (no thin middle `Api*` hop). Wave-10 (#4183): module bodies live
-/// in sibling libraries; this file is the stable forwarder barrel.
-abstract final class GameMapAreaStateLogic {
-  GameMapAreaStateLogic._();
-
-  static bool allowsFullTurnResolution(ct_models.Game game) =>
-      GameMapAreaStateLogicShell.allowsFullTurnResolution(game);
-
-  static const ({bool showIcon, bool enabled, bool hasExplorerUnits})
-  kHiddenExplorerInlineActionState =
-      GameMapAreaProvinceActionStates.kHiddenExplorerInlineActionState;
-  static const ({bool showIcon, bool enabled, bool hasBuilderUnits})
-  kHiddenBuilderInlineActionState =
-      GameMapAreaProvinceActionStates.kHiddenBuilderInlineActionState;
-  static const ({bool showIcon, bool enabled, bool hasEngineerUnits})
-  kHiddenEngineerInlineActionState =
-      GameMapAreaProvinceActionStates.kHiddenEngineerInlineActionState;
-  static const ({bool showIcon, bool enabled, bool hasMerchantUnits})
-  kHiddenMerchantInlineActionState =
-      GameMapAreaProvinceActionStates.kHiddenMerchantInlineActionState;
-
-  static int regionIndexFromWorldRegionId(String regionId) =>
-      GameMapAreaStateLogicShell.regionIndexFromWorldRegionId(regionId);
-
-  static ShellEntryAutoCenter? resolveShellEntryAutoCenter({
-    required ct_models.Game game,
-    required String? currentPlayerId,
-  }) =>
-      GameMapAreaStateLogicShell.resolveShellEntryAutoCenter(
-        game: game,
-        currentPlayerId: currentPlayerId,
-      );
-
-  static String translateWorkTargetTileKey({
-    required String tileKey,
-    required String workTarget,
-  }) =>
-      GameMapAreaStateLogicShell.translateWorkTargetTileKey(
-        tileKey: tileKey,
-        workTarget: workTarget,
-      );
-
-  static const Set<String> kCacheFirstWorkTargets =
-      GameMapAreaStateLogicWorkTargets.kCacheFirstWorkTargets;
-
-  static Set<String> filterCacheSelectionForRuntimeStaleTileConflicts({
-    required Set<String> cachedTileKeys,
-    required ct_models.Game game,
-    required ct_models.Orders currentOrders,
-    required String playerId,
-    required String selectedUnitId,
-    required String workTarget,
-  }) =>
-      GameMapAreaStateLogicWorkTargets
-          .filterCacheSelectionForRuntimeStaleTileConflicts(
-        cachedTileKeys: cachedTileKeys,
-        game: game,
-        currentOrders: currentOrders,
-        playerId: playerId,
-        selectedUnitId: selectedUnitId,
-        workTarget: workTarget,
-      );
-
-  static Set<String> resolveValidTileKeysForCivilianWorkSelection({
-    required String workTarget,
-    required PerPlayerWorkTargetSelectionCache workTargetSelectionCache,
-    required String humanPlayerId,
-    required String selectedUnitId,
-    required ct_models.Game game,
-    required ct_models.Orders currentOrders,
-    required PlayerView playerView,
-    required MapTopology topology,
-    required Map<String, TileMapResult>? tileMapByRegion,
-  }) =>
-      GameMapAreaStateLogicWorkTargets
-          .resolveValidTileKeysForCivilianWorkSelection(
-        workTarget: workTarget,
-        workTargetSelectionCache: workTargetSelectionCache,
-        humanPlayerId: humanPlayerId,
-        selectedUnitId: selectedUnitId,
-        game: game,
-        currentOrders: currentOrders,
-        playerView: playerView,
-        topology: topology,
-        tileMapByRegion: tileMapByRegion,
-      );
-
-  static ct_models.Orders addHumanWorkOrder({
-    required ct_models.Orders orders,
-    required String humanPlayerId,
-    required ct_models.WorkOrder workOrder,
-  }) =>
-      GameMapAreaStateLogicWorkTargets.addHumanWorkOrder(
-        orders: orders,
-        humanPlayerId: humanPlayerId,
-        workOrder: workOrder,
-      );
-
-  static String? selectionAfterWorkAssignment({
-    required String? currentSelectedCivilianTileKey,
-    required String assignedTileKey,
-  }) =>
-      GameMapAreaStateLogicWorkTargets.selectionAfterWorkAssignment(
-        currentSelectedCivilianTileKey: currentSelectedCivilianTileKey,
-        assignedTileKey: assignedTileKey,
-      );
-
-  static RegionMapViewData projectCivilianMarkersForHumanDraft({
-    required RegionMapViewData region,
-    required ct_models.Game game,
-    required ct_models.Orders orders,
-    required String humanPlayerId,
-    Set<String>? civilianMarkerOwnerIds,
-  }) =>
-      GameMapAreaStateLogicDraftProjection.projectCivilianMarkersForHumanDraft(
-        region: region,
-        game: game,
-        orders: orders,
-        humanPlayerId: humanPlayerId,
-        civilianMarkerOwnerIds: civilianMarkerOwnerIds,
-      );
-
-  static RegionMapViewData projectFleetMarkersForHumanDraft({
-    required RegionMapViewData region,
-    required ct_models.Game game,
-    required ct_models.Orders orders,
-    required String humanPlayerId,
-    required Map<String, TileMapResult> tileMapByRegion,
-    required Map<String, MapTopology> topologyByRegion,
-    required MapTopology combinedTopology,
-  }) =>
-      GameMapAreaStateLogicDraftProjection.projectFleetMarkersForHumanDraft(
-        region: region,
-        game: game,
-        orders: orders,
-        humanPlayerId: humanPlayerId,
-        tileMapByRegion: tileMapByRegion,
-        topologyByRegion: topologyByRegion,
-        combinedTopology: combinedTopology,
-      );
-
-  static RegionMapViewData projectHumanDraftMarkersForRegion({
-    required RegionMapViewData baseRegion,
-    required ct_models.Game game,
-    required ct_models.Orders orders,
-    required String humanPlayerId,
-    Map<String, TileMapResult>? tileMapByRegion,
-    Map<String, MapTopology>? topologyByRegion,
-    MapTopology? combinedTopology,
-    Set<String>? civilianMarkerOwnerIds,
-  }) =>
-      GameMapAreaStateLogicDraftProjection.projectHumanDraftMarkersForRegion(
-        baseRegion: baseRegion,
-        game: game,
-        orders: orders,
-        humanPlayerId: humanPlayerId,
-        tileMapByRegion: tileMapByRegion,
-        topologyByRegion: topologyByRegion,
-        combinedTopology: combinedTopology,
-        civilianMarkerOwnerIds: civilianMarkerOwnerIds,
-      );
-
-  static ({bool showIcon, bool enabled, bool hasExplorerUnits})
-  provinceProspectActionState({
-    required ct_models.Game game,
-    required String humanPlayerId,
-    required String selectedTileKey,
-    required PlayerView playerView,
-    required MapTopology? topology,
-    required ct_models.Orders currentOrders,
-    required Map<String, TileMapResult>? tileMapByRegion,
-  }) =>
-      GameMapAreaStateLogicProvinceActions.provinceProspectActionState(
-        game: game,
-        humanPlayerId: humanPlayerId,
-        selectedTileKey: selectedTileKey,
-        playerView: playerView,
-        topology: topology,
-        currentOrders: currentOrders,
-        tileMapByRegion: tileMapByRegion,
-      );
-
-  static Set<String> buildExploreEligibleTileKeyCache({
-    required ct_models.Game game,
-    required String humanPlayerId,
-    required PlayerView playerView,
-    required MapTopology topology,
-    required Map<String, TileMapResult>? tileMapByRegion,
-    required ct_models.Orders currentOrders,
-  }) =>
-      GameMapAreaStateLogicProvinceActions.buildExploreEligibleTileKeyCache(
-        game: game,
-        humanPlayerId: humanPlayerId,
-        playerView: playerView,
-        topology: topology,
-        tileMapByRegion: tileMapByRegion,
-        currentOrders: currentOrders,
-      );
-
-  static ({bool showIcon, bool enabled, bool hasExplorerUnits})
-  provinceExploreActionState({
-    required ct_models.Game game,
-    required String humanPlayerId,
-    required String selectedTileKey,
-    required RegionMapViewData selectedRegion,
-    PerPlayerWorkTargetSelectionCache? workTargetSelectionCache,
-    Set<String>? cachedExploreEligibleTileKeys,
-  }) =>
-      GameMapAreaStateLogicProvinceActions.provinceExploreActionState(
-        game: game,
-        humanPlayerId: humanPlayerId,
-        selectedTileKey: selectedTileKey,
-        selectedRegion: selectedRegion,
-        workTargetSelectionCache: workTargetSelectionCache,
-        cachedExploreEligibleTileKeys: cachedExploreEligibleTileKeys,
-      );
-
-  static ({bool showIcon, bool enabled, bool hasBuilderUnits})
-  provinceBuildImprovementActionState({
-    required ct_models.Game game,
-    required String humanPlayerId,
-    required String selectedTileKey,
-    required PlayerView playerView,
-    PerPlayerWorkTargetSelectionCache? workTargetSelectionCache,
-    MapTopology? topology,
-    ct_models.Orders currentOrders = const ct_models.Orders(),
-    Map<String, TileMapResult>? tileMapByRegion,
-  }) =>
-      GameMapAreaStateLogicProvinceActions.provinceBuildImprovementActionState(
-        game: game,
-        humanPlayerId: humanPlayerId,
-        selectedTileKey: selectedTileKey,
-        playerView: playerView,
-        workTargetSelectionCache: workTargetSelectionCache,
-        topology: topology,
-        currentOrders: currentOrders,
-        tileMapByRegion: tileMapByRegion,
-      );
-
-  static ({bool showIcon, bool enabled, bool hasEngineerUnits})
-  provinceBuildRoadActionState({
-    required ct_models.Game game,
-    required String humanPlayerId,
-    required String selectedTileKey,
-    required PlayerView playerView,
-    PerPlayerWorkTargetSelectionCache? workTargetSelectionCache,
-    MapTopology? topology,
-    ct_models.Orders currentOrders = const ct_models.Orders(),
-    Map<String, TileMapResult>? tileMapByRegion,
-  }) =>
-      GameMapAreaStateLogicProvinceActions.provinceBuildRoadActionState(
-        game: game,
-        humanPlayerId: humanPlayerId,
-        selectedTileKey: selectedTileKey,
-        playerView: playerView,
-        workTargetSelectionCache: workTargetSelectionCache,
-        topology: topology,
-        currentOrders: currentOrders,
-        tileMapByRegion: tileMapByRegion,
-      );
-
-  static ({bool showIcon, bool enabled, bool hasEngineerUnits})
-  provinceBuildFortActionState({
-    required ct_models.Game game,
-    required String humanPlayerId,
-    required String selectedTileKey,
-    required PlayerView playerView,
-    PerPlayerWorkTargetSelectionCache? workTargetSelectionCache,
-    MapTopology? topology,
-    ct_models.Orders currentOrders = const ct_models.Orders(),
-    Map<String, TileMapResult>? tileMapByRegion,
-  }) =>
-      GameMapAreaStateLogicProvinceActions.provinceBuildFortActionState(
-        game: game,
-        humanPlayerId: humanPlayerId,
-        selectedTileKey: selectedTileKey,
-        playerView: playerView,
-        workTargetSelectionCache: workTargetSelectionCache,
-        topology: topology,
-        currentOrders: currentOrders,
-        tileMapByRegion: tileMapByRegion,
-      );
-
-  static ({bool showIcon, bool enabled, bool hasEngineerUnits})
-  provinceBuildPortActionState({
-    required ct_models.Game game,
-    required String humanPlayerId,
-    required String selectedTileKey,
-    required PlayerView playerView,
-    PerPlayerWorkTargetSelectionCache? workTargetSelectionCache,
-    MapTopology? topology,
-    ct_models.Orders currentOrders = const ct_models.Orders(),
-    Map<String, TileMapResult>? tileMapByRegion,
-  }) =>
-      GameMapAreaStateLogicProvinceActions.provinceBuildPortActionState(
-        game: game,
-        humanPlayerId: humanPlayerId,
-        selectedTileKey: selectedTileKey,
-        playerView: playerView,
-        workTargetSelectionCache: workTargetSelectionCache,
-        topology: topology,
-        currentOrders: currentOrders,
-        tileMapByRegion: tileMapByRegion,
-      );
-
-  static ({bool showIcon, bool enabled, bool hasMerchantUnits})
-  provincePurchaseLandActionState({
-    required ct_models.Game game,
-    required String humanPlayerId,
-    required String selectedTileKey,
-    required PlayerView playerView,
-    PerPlayerWorkTargetSelectionCache? workTargetSelectionCache,
-    MapTopology? topology,
-    ct_models.Orders currentOrders = const ct_models.Orders(),
-    Map<String, TileMapResult>? tileMapByRegion,
-  }) =>
-      GameMapAreaStateLogicProvinceActions.provincePurchaseLandActionState(
-        game: game,
-        humanPlayerId: humanPlayerId,
-        selectedTileKey: selectedTileKey,
-        playerView: playerView,
-        workTargetSelectionCache: workTargetSelectionCache,
-        topology: topology,
-        currentOrders: currentOrders,
-        tileMapByRegion: tileMapByRegion,
-      );
-
-  static ({
-    bool showControl,
-    bool enabled,
-    bool hasBuilderUnits,
-    String? townTileKey,
-  })
-  provinceUpgradeTownActionState({
-    required ct_models.Game game,
-    required String humanPlayerId,
-    required String provinceId,
-    required PlayerView playerView,
-    PerPlayerWorkTargetSelectionCache? workTargetSelectionCache,
-    MapTopology? topology,
-    ct_models.Orders currentOrders = const ct_models.Orders(),
-    Map<String, TileMapResult>? tileMapByRegion,
-  }) =>
-      GameMapAreaStateLogicProvinceActions.provinceUpgradeTownActionState(
-        game: game,
-        humanPlayerId: humanPlayerId,
-        provinceId: provinceId,
-        playerView: playerView,
-        workTargetSelectionCache: workTargetSelectionCache,
-        topology: topology,
-        currentOrders: currentOrders,
-        tileMapByRegion: tileMapByRegion,
-      );
-}
