+++ app/test/features/game/flame/game_map_area_civilian_draft_projection_observe_test.dart
@@ -120,7 +120,7 @@ void main() {
           const orders = ct_models.Orders();
 
           final viaForwarder =
-              GameMapAreaStateLogic.projectCivilianMarkersForHumanDraft(
+              GameMapAreaStateLogicDraftProjection.projectCivilianMarkersForHumanDraft(
                 region: region,
                 game: game,
                 orders: orders,
