+++ app/test/game_map_area_state_logic_fleet_cross_region_test.dart
@@ -154,7 +154,7 @@ void main() {
         required Map<String, TileMapResult> tiles,
         required Map<String, MapTopology> topologies,
       }) {
-        return GameMapAreaStateLogic.projectFleetMarkersForHumanDraft(
+        return GameMapAreaStateLogicDraftProjection.projectFleetMarkersForHumanDraft(
           region: region,
           game: game,
           orders: orders,
