+++ app/test/game_map_area_state_logic_test_support.dart
@@ -19,7 +19,7 @@ import 'package:flutter_test/flutter_test.dart';
 /// Expected `provinceBuildImprovementActionState(...).enabled` per **pipeline
 /// contract A** (`SPEC/program/order-suggestions.md` Province Tile
 /// `Build improvement` shortcut enablement): same predicate as
-/// `GameMapAreaStateLogic.provinceBuildImprovementActionState` — any human
+/// `GameMapAreaStateLogicProvinceActions.provinceBuildImprovementActionState` — any human
 /// Builder whose allowed targets include `build_improvement` has
 /// `selectedTileKey` in `getValidWorkOrderTileKeysWithVisibility` for the same
 /// `(game, topology, view, orders, tileMap)`.
